Modify `io.js` to remove the use of `new` and `this`.
Looking at `infinispan.js` I noticed that the `io` package was being
used in an unusual way. [Here]
(https://github.com/infinispan/js-client/blob/master/lib/infinispan.js#L30)
the code is calling `new io.transport()`. This `transport` function is
also calling `new` [here]
(https://github.com/infinispan/js-client/blob/master/lib/io.js#L13-L15).
This is unnecessary, and potentially confusing. This commit addresses that.

In addition, @galderz, you mentioned that you were interested in keeping
the code as functional as possible, so in making the change above, I
have also eliminated the use of `Transport` as a constructor in favor of
it simply being a function. Doing that does not necessitate the removal
of `this`, but in general, I find that making assumptions about `this`
can lead to errors and in functional programming, it really shouldn't be
needed. So I've made changes to `io.transport` to eliminate `this` as
well.

Lastly, I added a basic .jshintrc and did a little code linting by
adding semi-colons.